Small comment

+++ index.cgi
@@ -963,7 +963,8 @@ sub postdata {
 # Helper to clear the cached files from the data dir.
 sub clearcachefiles {
   foreach my $pf ( glob($datadir."*") ) {
-    next if ( $pf =~ /\.data$/ );
+    next if ( $pf =~ /\.data$/ ); # .data files are the only really important ones
+    next if ( -d $pf ); # Skip subdirs, if we have such
     if ( $pf =~ /\/$username.*png/ ||   # All png files for this user
          -M $pf > 7 ) {  # And any file older than a week
       unlink ($pf)
